requirements.txt
Django==4.1
psycopg2-binary==2.9.9
djangorestframework==3.15.0
# psycopg2==2.9.9